requirements.txt
requests>=2.33
google-auth>=2.49
google-auth-oauthlib>=1.3
google-auth-httplib2>=0.3
google-api-python-client>=2.193
python-dotenv>=1.2
